common/allocators: Use 64 bit ints for all counts and sizes.
With the data plugin it is possible to create subpatterns which are
larger than a uint32_t.  This causes the plugin to silently overflow
when doing operations like '*'.  eg. This impossible operation worked
before:

$ ./nbdkit data '
           1 *2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2
             *2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2 ' \
           --run 'qemu-img info $nbd'
image: nbd://localhost:10809
file format: raw
virtual size: 1 TiB (1099511627776 bytes)

Simply changing all the functions to take uint64_t is sufficient.  The
data plugin is already passing 64 bit ints for counts, and the memory
plugin uses 32 bit ints which will be extended because the functions
are prototyped.

Thanks: Eric Blake
(cherry picked from commit 058be0a7410c3da0043b16df7240f71f99e68e51)

+++ common/allocators/malloc.c
@@ -149,7 +149,7 @@ m_alloc_set_size_hint (struct allocator *a, uint64_t size_hint)
 
 static int
 m_alloc_read (struct allocator *a, void *buf,
-              uint32_t count, uint64_t offset)
+              uint64_t count, uint64_t offset)
 {
   struct m_alloc *ma = (struct m_alloc *) a;
   ACQUIRE_RDLOCK_FOR_CURRENT_SCOPE (&ma->lock);
@@ -171,7 +171,7 @@ m_alloc_read (struct allocator *a, void *buf,
 
 static int
 m_alloc_write (struct allocator *a, const void *buf,
-               uint32_t count, uint64_t offset)
+               uint64_t count, uint64_t offset)
 {
   struct m_alloc *ma = (struct m_alloc *) a;
 
@@ -188,7 +188,7 @@ m_alloc_write (struct allocator *a, const void *buf,
 }
 
 static int
-m_alloc_fill (struct allocator *a, char c, uint32_t count, uint64_t offset)
+m_alloc_fill (struct allocator *a, char c, uint64_t count, uint64_t offset)
 {
   struct m_alloc *ma = (struct m_alloc *) a;
 
@@ -202,7 +202,7 @@ m_alloc_fill (struct allocator *a, char c, uint32_t count, uint64_t offset)
 }
 
 static int
-m_alloc_zero (struct allocator *a, uint32_t count, uint64_t offset)
+m_alloc_zero (struct allocator *a, uint64_t count, uint64_t offset)
 {
   struct m_alloc *ma = (struct m_alloc *) a;
   ACQUIRE_RDLOCK_FOR_CURRENT_SCOPE (&ma->lock);
@@ -222,7 +222,7 @@ m_alloc_zero (struct allocator *a, uint32_t count, uint64_t offset)
 
 static int
 m_alloc_blit (struct allocator *a1, struct allocator *a2,
-              uint32_t count, uint64_t offset1, uint64_t offset2)
+              uint64_t count, uint64_t offset1, uint64_t offset2)
 {
   struct m_alloc *ma2 = (struct m_alloc *) a2;
 
@@ -239,7 +239,7 @@ m_alloc_blit (struct allocator *a1, struct allocator *a2,
 
 static int
 m_alloc_extents (struct allocator *a,
-                 uint32_t count, uint64_t offset,
+                 uint64_t count, uint64_t offset,
                  struct nbdkit_extents *extents)
 {
   /* Always fully allocated.  XXX In theory we could detect zeroes
